+++ src/app.ts
@@ -1,11 +1,27 @@
+/*
+ Copyright (C) 2022-present Metahkg Contributors
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU Affero General Public License as
+ published by the Free Software Foundation, either version 3 of the
+ License, or (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU Affero General Public License for more details.
+
+ You should have received a copy of the GNU Affero General Public License
+ along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 import dotenv from "dotenv";
 import routes from "./routes";
-import Fastify from "fastify";
+import Fastify, { FastifyRequest } from "fastify";
 import { client, domain } from "./lib/common";
 import { setup } from "./mongo/setupMongo";
 import { agenda } from "./lib/agenda";
 import refreshToken from "./plugins/refreshToken";
-import updateToken from "./plugins/updateToken";
 import multipart from "@fastify/multipart";
 import fastifyRateLimit from "@fastify/rate-limit";
 import fastifyCors from "@fastify/cors";
@@ -66,11 +82,14 @@ export default async function MetahkgServer() {
 
     fastify.register(multipart);
 
-    fastify.register(fastifyRateLimit, {
+    await fastify.register(fastifyRateLimit, {
         global: true,
         max: 200,
         ban: 50,
         timeWindow: 1000 * 30,
+        keyGenerator: (req: FastifyRequest) => {
+            return sha256(req.ip);
+        },
     });
 
     fastify.register(fastifyJwt, {
@@ -94,7 +113,6 @@ export default async function MetahkgServer() {
     });
 
     fastify.addHook("onRequest", authenticate);
-    fastify.addHook("onRequest", updateToken);
     fastify.addHook("onRequest", refreshToken);
     // re-verify after updateToken and refreshToken
     fastify.addHook("onRequest", authenticate);