+++ src/lib/notify.ts
@@ -1,2 +1,0 @@
-// TODO: notify users on reply
-// export async function notify(threadId: number, commentId: number) {}