+++ src/lib/sessions/createSession.ts
@@ -1,15 +1,34 @@
+/*
+ Copyright (C) 2022-present Metahkg Contributors
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU Affero General Public License as
+ published by the Free Software Foundation, either version 3 of the
+ License, or (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU Affero General Public License for more details.
+
+ You should have received a copy of the GNU Affero General Public License
+ along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 import { randomBytes } from "crypto";
 import { usersCl } from "../common";
 import { sha256 } from "../sha256";
 import { createDecoder } from "fast-jwt";
 import { jwtTokenType } from "../../types/jwt";
+import User from "../../models/user";
 
 export async function createSession(
     userId: number,
     token: string,
     userAgent: string,
     ip: string,
-    sameIp?: boolean
+    sameIp?: boolean,
+    createdAt: Date = new Date()
 ) {
     const decode = createDecoder();
     const exp = (decode(token) as jwtTokenType)?.exp * 1000;
@@ -18,7 +37,7 @@ export async function createSession(
     const session = {
         id: randomBytes(15).toString("hex"),
         token: sha256(token),
-        createdAt: new Date(),
+        createdAt,
         exp: new Date(exp),
         userAgent,
         ip: sha256(ip),
@@ -26,10 +45,10 @@ export async function createSession(
     };
 
     while (
-        await usersCl.findOne({
+        (await usersCl.findOne({
             id: userId,
             sessions: { $elemMatch: { id: session.id } },
-        })
+        })) as User
     ) {
         session.id = randomBytes(15).toString("hex");
     }