+++ src/routes/categories/threads.ts
@@ -1,9 +1,27 @@
+/*
+ Copyright (C) 2022-present Metahkg Contributors
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU Affero General Public License as
+ published by the Free Software Foundation, either version 3 of the
+ License, or (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU Affero General Public License for more details.
+
+ You should have received a copy of the GNU Affero General Public License
+ along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 import { categoryCl, threadCl } from "../../lib/common";
 import { hiddencats as gethiddencats } from "../../lib/hiddencats";
 import { Static, Type } from "@sinclair/typebox";
 import Thread from "../../models/thread";
 import { FastifyInstance, FastifyPluginOptions, FastifyRequest } from "fastify";
 import regex from "../../lib/regex";
+import Category from "../../models/category";
 
 export default (
     fastify: FastifyInstance,
@@ -48,7 +66,7 @@ export default (
             if (!req.user && hiddenCats.includes(category))
                 return res.code(403).send({ statusCode: 403, error: "Forbidden." });
 
-            if (!(await categoryCl.findOne({ id: category })))
+            if (!((await categoryCl.findOne({ id: category })) as Category))
                 return res
                     .code(404)
                     .send({ statusCode: 404, error: "Category not found." });