+++ src/routes/me/following.ts
@@ -1,3 +1,20 @@
+/*
+ Copyright (C) 2022-present Metahkg Contributors
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU Affero General Public License as
+ published by the Free Software Foundation, either version 3 of the
+ License, or (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU Affero General Public License for more details.
+
+ You should have received a copy of the GNU Affero General Public License
+ along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 import { FastifyInstance, FastifyPluginOptions } from "fastify";
 import { usersCl } from "../../lib/common";
 import User, { FollowedUser } from "../../models/user";
@@ -12,10 +29,10 @@ export default function (
         if (!user) return res.code(401).send({ statusCode: 401, error: "Unauthorized." });
 
         const following = ((
-            await usersCl.findOne(
+            (await usersCl.findOne(
                 { id: user.id },
                 { projection: { _id: 0, following: 1 } }
-            )
+            )) as User
         )?.following || []) as FollowedUser[];
 
         const usersFollowed = (await usersCl