+++ src/routes/threads/thread/actions/delete.ts
@@ -1,3 +1,20 @@
+/*
+ Copyright (C) 2022-present Metahkg Contributors
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU Affero General Public License as
+ published by the Free Software Foundation, either version 3 of the
+ License, or (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU Affero General Public License for more details.
+
+ You should have received a copy of the GNU Affero General Public License
+ along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 import { Static, Type } from "@sinclair/typebox";
 import { FastifyInstance, FastifyPluginOptions, FastifyRequest } from "fastify";
 import { removedCl, threadCl } from "../../../../lib/common";
@@ -5,6 +22,8 @@ import { removedCl, threadCl } from "../../../../lib/common";
 import regex from "../../../../lib/regex";
 import RequireAdmin from "../../../../plugins/requireAdmin";
 import { ReasonSchemaAdmin } from "../../../../lib/schemas";
+import Thread from "../../../../models/thread";
+import { objectFilter } from "../../../../lib/objectFilter";
 
 export default function (
     fastify: FastifyInstance,
@@ -34,9 +53,14 @@ export default function (
         ) => {
             const id = Number(req.params.id);
             const { reason } = req.body;
-            const admin = req.user;
+            const admin = objectFilter(req.user, (key: string) =>
+                ["id", "name", "sex", "role"].includes(key)
+            );
 
-            const thread = await threadCl.findOne({ id }, { projection: { _id: 0 } });
+            const thread = (await threadCl.findOne(
+                { id },
+                { projection: { _id: 0 } }
+            )) as Thread;
             if (!thread)
                 return res
                     .code(404)