+++ src/routes/threads/thread/actions/edit.ts
@@ -1,3 +1,20 @@
+/*
+ Copyright (C) 2022-present Metahkg Contributors
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU Affero General Public License as
+ published by the Free Software Foundation, either version 3 of the
+ License, or (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU Affero General Public License for more details.
+
+ You should have received a copy of the GNU Affero General Public License
+ along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 import { Static, Type } from "@sinclair/typebox";
 import { FastifyInstance, FastifyPluginOptions, FastifyRequest } from "fastify";
 import { categoryCl, threadCl } from "../../../../lib/common";
@@ -6,6 +23,8 @@ import regex from "../../../../lib/regex";
 import checkThread from "../../../../plugins/checkThread";
 import RequireAdmin from "../../../../plugins/requireAdmin";
 import { IntegerSchema, ReasonSchemaAdmin, TitleSchema } from "../../../../lib/schemas";
+import Category from "../../../../models/category";
+import { objectFilter } from "../../../../lib/objectFilter";
 
 export default function (
     fastify: FastifyInstance,
@@ -44,7 +63,7 @@ export default function (
 
             const { category, title, reason } = req.body;
 
-            if (category && !(await categoryCl.findOne({ id: category })))
+            if (category && !((await categoryCl.findOne({ id: category })) as Category))
                 return res
                     .code(404)
                     .send({ statusCode: 404, error: "Category not found" });
@@ -58,7 +77,9 @@ export default function (
                     },
                     $push: {
                         "admin.edits": {
-                            admin: user,
+                            admin: objectFilter(user, (key: string) =>
+                                ["id", "name", "sex", "role"].includes(key)
+                            ),
                             reason,
                             date: new Date(),
                         },