+++ src/routes/threads/thread/actions/star.ts
@@ -1,8 +1,26 @@
+/*
+ Copyright (C) 2022-present Metahkg Contributors
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU Affero General Public License as
+ published by the Free Software Foundation, either version 3 of the
+ License, or (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU Affero General Public License for more details.
+
+ You should have received a copy of the GNU Affero General Public License
+ along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 import { Static, Type } from "@sinclair/typebox";
 import { FastifyInstance, FastifyPluginOptions, FastifyRequest } from "fastify";
 import { threadCl, usersCl } from "../../../../lib/common";
 
 import regex from "../../../../lib/regex";
+import Thread from "../../../../models/thread";
 
 export default function (
     fastify: FastifyInstance,
@@ -23,7 +41,7 @@ export default function (
 
             const threadId = Number(req.params.id);
 
-            if (!(await threadCl.findOne({ id: threadId })))
+            if (!((await threadCl.findOne({ id: threadId })) as Thread))
                 return res
                     .code(404)
                     .send({ statusCode: 404, error: "Thread not found." });