+++ src/routes/threads/thread/comments/emotion/users.ts
@@ -1,8 +1,26 @@
+/*
+ Copyright (C) 2022-present Metahkg Contributors
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU Affero General Public License as
+ published by the Free Software Foundation, either version 3 of the
+ License, or (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU Affero General Public License for more details.
+
+ You should have received a copy of the GNU Affero General Public License
+ along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 import { Static, Type } from "@sinclair/typebox";
 import { FastifyInstance, FastifyPluginOptions, FastifyRequest } from "fastify";
 import { threadCl, usersCl } from "../../../../../lib/common";
 import regex from "../../../../../lib/regex";
 import { Emotion } from "../../../../../models/thread";
+import User from "../../../../../models/user";
 
 export default function (
     fastify: FastifyInstance,
@@ -61,7 +79,7 @@ export default function (
 
             if (!emotions?.length) return res.send([]);
 
-            const users = await usersCl
+            const users = (await usersCl
                 .find({
                     id: { $in: emotions.map((x) => x.user) },
                 })
@@ -72,7 +90,7 @@ export default function (
                     sex: 1,
                     role: 1,
                 })
-                .toArray();
+                .toArray()) as User[];
 
             res.send(users);
         }