+++ src/routes/threads/thread/comments/vote.ts
@@ -1,3 +1,20 @@
+/*
+ Copyright (C) 2022-present Metahkg Contributors
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU Affero General Public License as
+ published by the Free Software Foundation, either version 3 of the
+ License, or (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU Affero General Public License for more details.
+
+ You should have received a copy of the GNU Affero General Public License
+ along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 import { domain, threadCl, votesCl } from "../../../../lib/common";
 import { Type, Static } from "@sinclair/typebox";
 
@@ -63,7 +80,7 @@ export default (
             const votes = (await votesCl.findOne({ id: user.id })) as Votes;
 
             if (!votes) {
-                await votesCl.insertOne({ id: user.id });
+                await votesCl.insertOne(<Votes>{ id: user.id });
             } else if (votes?.[threadId]?.find((i) => i.cid === commentId)) {
                 return res
                     .code(429)