+++ src/routes/users/login.ts
@@ -1,13 +1,38 @@
+/*
+ Copyright (C) 2022-present Metahkg Contributors
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU Affero General Public License as
+ published by the Free Software Foundation, either version 3 of the
+ License, or (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU Affero General Public License for more details.
+
+ You should have received a copy of the GNU Affero General Public License
+ along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 import dotenv from "dotenv";
-import { usersCl, verificationCl } from "../../lib/common";
+import { RecaptchaSecret, usersCl, verificationCl } from "../../lib/common";
 import bcrypt from "bcrypt";
 import { Static, Type } from "@sinclair/typebox";
 import { createToken } from "../../lib/auth/createToken";
 import User from "../../models/user";
 import { FastifyInstance, FastifyPluginOptions, FastifyRequest } from "fastify";
 import { createSession } from "../../lib/sessions/createSession";
-import { EmailSchema, PasswordSchema, UserNameSchema } from "../../lib/schemas";
+import {
+    EmailSchema,
+    PasswordSchema,
+    RTokenSchema,
+    UserNameSchema,
+} from "../../lib/schemas";
 import { sha256 } from "../../lib/sha256";
+import { Verification } from "../../models/verification";
+import { verifyCaptcha } from "../../lib/recaptcha";
+import { RateLimitOptions } from "@fastify/rate-limit";
 
 dotenv.config();
 
@@ -22,6 +47,7 @@ export default (
             // check if password is a sha256 hash
             password: PasswordSchema,
             sameIp: Type.Optional(Type.Boolean()),
+            rtoken: RTokenSchema,
         },
         { additionalProperties: false }
     );
@@ -29,26 +55,34 @@ export default (
     fastify.post(
         "/login",
         {
-            preHandler: fastify.rateLimit({
-                max: 5,
-                ban: 5,
-                timeWindow: 1000 * 60 * 5,
-            }),
+            config: {
+                rateLimit: <RateLimitOptions>{
+                    max: 5,
+                    ban: 5,
+                    timeWindow: 1000 * 60 * 5,
+                },
+            },
             schema: { body: schema },
         },
         async (req: FastifyRequest<{ Body: Static<typeof schema> }>, res) => {
-            const { name, password, sameIp } = req.body;
+            const { name, password, sameIp, rtoken } = req.body;
+
+            if (!(await verifyCaptcha(RecaptchaSecret, rtoken)))
+                return res
+                    .code(429)
+                    .send({ statusCode: 429, error: "Recaptcha token invalid." });
 
             const user = (await usersCl.findOne({
                 $or: [{ name }, { email: sha256(name) }],
             })) as User;
 
             if (!user) {
-                const verifyUser = await verificationCl.findOne({
+                const verifyUser = (await verificationCl.findOne({
+                    type: "register",
                     $or: [{ name }, { email: sha256(name) }],
-                });
+                })) as Verification & { type: "register" };
 
-                if (verifyUser && (await bcrypt.compare(password, verifyUser.pwd)))
+                if (verifyUser && (await bcrypt.compare(password, verifyUser.password)))
                     return res
                         .code(409)
                         .send({ statusCode: 409, error: "Please verify your email." });