+++ src/routes/users/user/actions/ban.ts
@@ -1,3 +1,20 @@
+/*
+ Copyright (C) 2022-present Metahkg Contributors
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU Affero General Public License as
+ published by the Free Software Foundation, either version 3 of the
+ License, or (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU Affero General Public License for more details.
+
+ You should have received a copy of the GNU Affero General Public License
+ along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 import { Static, Type } from "@sinclair/typebox";
 import { FastifyInstance, FastifyPluginOptions, FastifyRequest } from "fastify";
 import { usersCl } from "../../../../lib/common";
@@ -7,6 +24,7 @@ import regex from "../../../../lib/regex";
 import User from "../../../../models/user";
 import RequireAdmin from "../../../../plugins/requireAdmin";
 import { ReasonSchemaAdmin, DateSchema } from "../../../../lib/schemas";
+import { objectFilter } from "../../../../lib/objectFilter";
 
 export default function (
     fastify: FastifyInstance,
@@ -36,7 +54,9 @@ export default function (
             res
         ) => {
             const id = Number(req.params.id);
-            const admin = req.user;
+            const admin = objectFilter(req.user, (key: string) =>
+                ["id", "name", "sex", "role"].includes(key)
+            );
             const { reason, exp } = req.body;
 
             const reqUser = (await usersCl.findOne({ id })) as User;