+++ src/routes/users/user/actions/edit.ts
@@ -1,3 +1,20 @@
+/*
+ Copyright (C) 2022-present Metahkg Contributors
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU Affero General Public License as
+ published by the Free Software Foundation, either version 3 of the
+ License, or (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU Affero General Public License for more details.
+
+ You should have received a copy of the GNU Affero General Public License
+ along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 import { Static, Type } from "@sinclair/typebox";
 import { usersCl } from "../../../../lib/common";
 
@@ -5,7 +22,7 @@ import { createToken } from "../../../../lib/auth/createToken";
 import { FastifyInstance, FastifyPluginOptions, FastifyRequest } from "fastify";
 import regex from "../../../../lib/regex";
 import EmailValidator from "email-validator";
-import { userSex } from "../../../../models/user";
+import User, { userSex } from "../../../../models/user";
 import { updateSessionByToken } from "../../../../lib/sessions/updateSession";
 import { SexSchema, UserNameSchema } from "../../../../lib/schemas";
 
@@ -42,7 +59,7 @@ export default (
 
             const { name, sex } = req.body as { name?: string; sex?: userSex };
 
-            if (name && name !== user.name && (await usersCl.findOne({ name })))
+            if (name && name !== user.name && ((await usersCl.findOne({ name })) as User))
                 return res
                     .code(409)
                     .send({ statusCode: 409, error: "Name already taken." });